requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
